@@ -9,12 +9,13 @@
     "description": "PayPlug payment plugin for Sylius applications.",
     "license": "MIT",
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.0",
         "ext-json": "*",
         "giggsey/libphonenumber-for-php": "^8.12",
         "payplug/payplug-php": "^3.1",
-        "sylius/refund-plugin": "^1.0.0",
-        "sylius/sylius": "^1.9.0 || ^1.10.0 || ^1.11.0",
+        "php-http/message-factory": "^1.1",
+        "sylius/refund-plugin": "^1.4",
+        "sylius/sylius": "^1.11.0",
         "symfony/asset": "^4.4|^5.0|^6.0",
         "symfony/lock": "^4.4|^5.0|^6.0",
         "symfony/validator": "^4.4|^5.0|^6.0"
